Redundant deps eliminated

+++ src/processor/processor.rs
@@ -10,13 +10,13 @@ pub fn overlay_images(logo_image: DynamicImage, main_image: DynamicImage) -> Dyn
 
     let margin = 10;
 
-    let logo_position_x = main_width - logo_width - margin;
-    let logo_position_y = main_height - logo_height - margin;
+    let logo_position_x = (main_width - logo_width - margin) as i64;
+    let logo_position_y = (main_height - logo_height - margin) as i64;
 
     let resized_logo = logo_image.resize_exact(
         logo_width,
         logo_height,
-        imageops::FilterType::Lanczos3, 
+        imageops::FilterType::Lanczos3,
     );
 
     let logo_rgba = resized_logo.to_rgba8();
